requirements.txt
pyjwt[crypto]
python-binance
requests
pytelegrambotapi
keyboa